Today progress:
I've modified the Client code to be more small and elegant. Also, now the client waits from a Trip and support "being told to move by the Server". Lior fixed some bugs and logic in the client. The clientInput class is now unnecessary and will be deleted.
Another big change me and lior did is the RemoteDriver class, which is a wrapper(not really, RemoteDriver is a Driver, with didn't wrap it because we only needed to modify 3-4 small methods and its dumb to wrap every method of driver just for that) of Driver that adds socket support. With the RemoteDriver the TaxiCenter(which is a complex class) can contact clients without being aware of it. This will be great when will use TCP and every RemoteDriver will actually have a different Socket.
Lastly, The Server have changed to support some of the new game flow logic and bug fixed.